samples: shell_module: move login cmds into a separate file
Was getting compilation error due to `CONFIG_SHELL_CMD_ROOT`
not being defined whe building the sample under certain
configurations.

The `CONFIG_SHELL_CMD_ROOT` is defined and used only when
`CONFIG_SHELL_START_OBSCURED` is enabled, we can simply move
it to another file and rely on CMakeLists.txt to compile.

Run the `login_init` with `SYS_INIT` so that the `main()`
do not need to have knowledge of its existence.
